refactor: move modules to folder
