+++ src/infra/http/nest/modules/associate/associate.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 
 import { PrismaService } from '@/infra/repositories/prisma/prisma.service';
-// import { AssociateService } from '@/infra/http/nest/associate/associate.service';
-import { AssociateController } from '@/infra/http/nest/associate/associate.controller';
+import { AssociateController } from '@/infra/http/nest/modules/associate/associate.controller';
 import { CreateAssociateService } from '@/domains/associate/services/create-associate.service';
-import { PrismaAssociateRepository } from '@/infra/repositories/prisma/domains/prisma-asssociate.repository';
+import { PrismaAssociateRepository } from '@/infra/repositories/prisma/domains/prisma-associate.repository';
 import { UpdateAssociateService } from '@/domains/associate/services/update-associate.service';
 import { ListAssociateService } from '@/domains/associate/services/list-associate.service';
 import { FindAssociateByIdService } from '@/domains/associate/services/find-associate-by-id.service';
@@ -13,13 +12,6 @@ import { DeleteAssociateByIdService } from '@/domains/associate/services/delete-
 @Module({
   controllers: [AssociateController],
   providers: [
-    // {
-    //   provide: AssociateService,
-    //   useFactory: (prisma) => new AssociateService(prisma),
-    //   inject: [PrismaService],
-    // },
-    // PrismaService,
-
     {
       provide: CreateAssociateService,
       useFactory: (repository) => new CreateAssociateService(repository),
