+++ src/infra/http/nest/modules/payment/payment.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PrismaService } from '@/infra/repositories/prisma/prisma.service';
-import { PaymentService } from '@/infra/http/nest/payment/payment.service';
-import { PaymentController } from '@/infra/http/nest/payment/payment.controller';
+import { PaymentService } from '@/infra/http/nest/modules/payment/payment.service';
+import { PaymentController } from '@/infra/http/nest/modules/payment/payment.controller';
 
 @Module({
   controllers: [PaymentController],
